backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
python-multipart
python-dotenv
google-genai
pdfplumber
pypdf
